Fixing auth methods and wallet loading
