+++ helpers/utils.js
@@ -55,9 +55,9 @@ export const showMessageDialog = (message, msgType) => {
 export const showErrorMessage = (message) => {
   showMessage({
     duration: 5000,
-    message: "Error",
+    message: 'Error',
     description: message,
-    type: "danger",
+    type: 'danger',
     titleStyle: styles.errorText,
   });
 };
@@ -65,9 +65,9 @@ export const showErrorMessage = (message) => {
 export const showSuccessMessage = (message) => {
   showMessage({
     duration: 5000,
-    message: "Success",
+    message: 'Success',
     description: message,
-    type: "success",
+    type: 'success',
     titleStyle: styles.errorText,
   });
 };
